normalize extra metrics to gigagas
